cmd/snap-confine: remove setuid calls from cgroup init code
Calls to sc_set_effective_identity() are not needed in the cgroup
initialization code (both v1 and v2), because this code is executed
either in the context of snap-device-helper, which is started by udev
and therefore runs as root, or in snap-confine before switching to the
unprivileged user.

This commit introduces a little change, though: the group ID of the
leaf files created inside a v1 cgroup will be that of the user, since
our effective group ID is not 0 (snap-confine has only the root setuid
bit, not the setgid one). However, all these files are not writable for
the group, and are readable by everybody. So it seems fine to relax the
spread test to accept this.

The code paths in snap-confine are:

main()
  -> sc_setup_device_cgroup()
    -> sc_device_cgroup_new()
      -> _sc_cgroup_v1_init()  # for v1
      -> _sc_cgroup_v2_init()  # for v2
    -> sc_device_cgroup_attach_pid()
      -> _sc_cgroup_v2_attach_pid()  # for v2
        -> _sc_cgroup_v2_attach_pid_bpf()