interfaces/modem-manager: Only generate DBus plug policy on Core
Classic systems provide their own ModemManager along with it's own DBus
policy - if we also generate DBus policy for ModemManager when a snap plugs
this interface, this will likely conflict with the DBus policy from the
classic system itself and cause unexpected results. As such, only generate
DBus policy for ModemManager when a snap plugs this interface on
non-classic systems (ie. Ubuntu Core etc).

See
https://forum.snapcraft.io/t/dbus-access-denied-while-using-modem-manager-plug-in-snap/29037/12
for background and discussion.

Signed-off-by: Alex Murray <alex.murray@canonical.com>