o/ifacestate: warn if the snapd.apparmor service is disabled
If the snapd.apparmor service is not started, the AppArmor profiles of
snap-confine and snap-update-ns might not be loaded. If that happens, no
snap applications can start, as the calls to aa_change_onexec() will
fail.

We might want to take more drastic measures when this happens (like,
re-enabling the service ourselves?), but while we ponder the decision
emitting a warning seems like a harmless yet useful behavioural change.

See also: https://bugs.launchpad.net/snapd/+bug/1806135