tests: add nested test variant that adds 4k sector size
This test would have caught
https://bugs.launchpad.net/snappy/+bug/1968356 where UC22 does
not install on devices that have 4k physical sector sizes.

The downside of course is that it makes the nested tests slower.
Alternatively we could simply unconditionally set all basic tests
to 4k and rely on the other tests to catch possible regressions
with physical sector size of 512 (which is the default).