cmd/snap-confine: make detection of stale namespaces more robust
The changes introduced with the previous commit would work also in the
case that /usr was itself a tmpfs (created as a writable mimic because
of a layout entry placed directly under /usr), but only because of the
existance of the bind mount of /usr onto itself, created by the code
that sets up writable mimics.
It was a bit too subtle and could have broken if this bind-mount were
not done, so it's better not to rely on that.

With this change here, instead, we rework the device detection code in a
way that more reliably identifies whether the base snap has changed.