tests: force version 2.48.3 on xenial ESM
When the "upgrade-from-release" tests runs on 16.04 we know
exactly what version to expect (2.48.3) because 16.04 is in ESM
mode and the archive will never change unless one enables ESM.

So force installing this version to ensure that even if some
test or something enables the ppa:snappy-dev/image PPA we always
get the right version.

We should probably also write an invariant test that ensures that
the sources.list does not change between runs but that would be
a followup.