features: disable refresh-app-awarness by default again
* features: disable refresh-app-awarness by default again

We enabled the `experimental.refresh-app-awareness` option right
before the 22.04 release [1]. However in hindsight this move was
premature. The UX is not quite ready, the refresh is stopped
but the experience should be that once the app is closed the
update gets applied which is not implemented yet. There are
also design issues with e.g. LXD as outlined in LP:#1978005
the need time and design to resolve. Another reason was the
automatic ~/snap -> ~/Snap folder migration. But this is
done differently now.

For the above reasons this commit changes the default again
to `refresh-app-awarness=false` until there is time to
fix the UX and the LXD issues.

[1] https://github.com/snapcore/snapd/pull/11566

* hookstate: disable refresh-app-awareness in gateAutoRefreshHookSuite.SetUpTest() - not strictly needed but when we enable refresh-app-awareness again it will help