tests: tweak the "tests/nested/manual/connections" test
The `tests/nested/manual/connections` is failing sometimes and
it's unclear why. When it fails it appears that the `connections`
snap is not installed which is strange given that it's clearly
needed in the model - as a consequence there are no connections
so the test fails.

To untangle this a bit this commit makes some changes to the test:
- rename to gadget-connections
- rename test snap from just "connections" to "tests-snapd-connections"
- rename the model from "pc" to "nested-20-amd64-test-snapd-connections"
- precondition checks that ensure we have the right model
- precondition checks that we have the expected test-snapd-connections
  snap
- move image building into "execute:" to ensure that on failure we get
  the shell log of how the image was build (this is lost if "prepare"
  bulds the image but the error happens in "execute").