tests: remove holds before auto-refreshing in tests
Tests relying on auto-refreshing triggered them by setting the
'last-refresh' to distant time which would then exceed the maximum
postponement limit and trigger an auto-refresh despite any holds. Since
refresh can now be held indefinitely and the system starts up with a 2h
hold, we need to remove this hold in order for the auto-refresh to trigger.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>