Reload ssh.service, not sshd.service
In Debian and Ubuntu, sshd.service is an alias for ssh.service and always has
been.  Reload the service by its proper name in our test suite.

This is necessary because ssh is switching to socket activation in later
Ubuntu releases, which means the ssh.service is no longer enabled (in favor
of ssh.socket) and therefore the sshd.service alias is not registered.