devicestate: return an error in checkEncryption() if KernelInfo fails
The current code in DeviceManager.checkEncryption() will not return
an error if getting the KernelInfo fails. It was done this way
because historically we did not mock enough in many places in the
tests.

Now the tests need only a small update to do the proper checking
and error handling here. This commits adds the missing mocks and
updates the code.