sandbox/apparmor: Add test for apparmor_parser exit 0 with error
Older versions of apparmor_parser may actually exit with 0 status but fail to
parse the profile - this can be detected by looking for "parser error" in the
output from apparmor_parser. The apparmor sandbox code now handles this
correctly, so add a test to ensure we are testing this behaviour going forward.

Signed-off-by: Alex Murray <alex.murray@canonical.com>