devicestate: update testInstallGadgetNoSave()
The testInstallGadgetNoSave() was testing if a dangerous model
can be installed on a gadget that does not have "ubuntu-save".

The old code was failing here because `checkEncryption()` did
not consider the gadget so it returned that encryption should
be used but that is not compatible with the gadget.

However the new code is smarter and warns about incompatible
gadgets and switches the encryption off if it's not required.

So the tests needs updating to a model that requires encryption.