Remove NESTED_IMAGE_ID from nested manual tests
As now the image name is calculated automatically based on the test name
and variant for manual nested tests, it is not required anymore to setup
that in the test environment.

This was causing problems in some tests, i.e. the core20-4k-sector-size
test was setting just 1 NESTED_IMAGE_ID but it had 2 variants, so when
both variants were executed in the same host, both variants were using
the same image and the second execution was failing because of that.