store: send Snap-Device-Location header with cloud information
The Store currently gets network location hints from a device using the
download API's Snap-CDN header. But it is useful for Store operations to
have that information even in e.g. refresh, so send the same cloud
information to all APIs in a new header.

The Store might eventually use that if Snap-CDN isn't present, leaving
Snap-CDN to handle only the "none" case, but for now leave that alone
for compatibility.