tests: ensure that fakeinstaller put the seed into the right place
When the fakeinstaller installs a system it will copy the selected
seed on the target system. However for some pre-generated images
there is a seed directory already on the target. In this case
the fakeinstaller needs to remove it and make sure that the
seed selected for install gets put into that place.