tests: use dbus-daemon instead of dbus-launch
This test has not been running on all Ubuntu machines since dbus-launch
was removed from the spread machines (commit 62f1ad5d). So, stop using
dbus-launch in the test and directly start dbus-daemon.

Also, fix a test failure occurring on Arch, where the new D-Bus session
is created with a regular socket on /tmp, which is not found from inside
the snap:

    + MATCH 'hello world'
    + test-snapd-dbus-service-client.session
    Failed to open connection to "session" message bus: Failed to connect to socket /tmp/dbus-zIatB2DeHj: No such file or directory

Passing the `--address` option to force the dbus-daemon to listen on an
abstract socket fixes this issue.