Update labeler configuration
the any: [] just works when all: [] is used, otherwise weeeee need to
use a list that means the same than any.

I tested that in the snapd-testing-tools project

Then "sync-labels" by default is false, it means that when a file is
removed from the pr, the label is not removed. But if we manually remove
the label, then when the workflow is executed again, la label will be
applied again following the rules defined.