tests/main/snapd-apparmor: make log parsing more robust
This line of the test would be succeeding even if no additional output
was emitted by snapd-apparmor, if th expected line was already present
in the logs (maybe written at boot time or by some other test).

Use the `journal-state` program to read the logs, since it uses a cursor
to limit the search to the logs emitted during the current test
execution.

It also comes with a retry mechanism that can help to make the test more
robust.