gadget: rename some OnDiskStructure fields
Rename some OnDiskStructure fields so they more clearly reflect that
they are used for the partition and not for the content of the LUKS
container (in case it exists). For instance, in the encryption case
the partition filesystem will be crypto_LUKS but the LUKS
container (mapped device) itself will have an ext4 filesystem.