hookstate: log ignored hook errors also to the journal
This commit adds logging of ignored errors to the journal. The
rational is that if a non-ephemeral hook has `IgnoreError` set
then the error is only logged to the task right now. But the
task will be in `Done` state and not in `Error` state so it is
hard for users to find problems.

This fixes the issue that during firstboot seeding an error
in the gadget.yaml defaults configuration is ignored but it's
not easy for the user to find the reason for this.