tests: fix broken snap-run-symlink-error test
The return code being captured was that of the "echo" and not the
command being tested. This wasn't noticed because although the
error code check echoed an error message, it didn't fail the test.
This was presumably because the command's output would also fail if
the error code was wrong but that wasn't the case because the code was
taken from the wrong command.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>