Cleanup the tests environment to avoid dirty tests directory
Some tests are failing because there are some directories which contain
snaps before the execution.

This is mostly for the first test executed, which contained many snaps
and directories created while the test suite is prepared.

The change cleans correctly the suite preaparation and moved the extra
tests directory to /tmp for the same raeson.