store: fix cache miss when snap was already in download path
If a snap was already in the target path, the cache would try to link
the cached file to that path, fail and return an os.ErrExist. Since the
store code was interpreting any error from the cache as a cache miss,
we would re-download it. In practice this didn't happen often because
when installing/refreshing a snap the file is eventually deleted from
that location so subsequent downloads wouldn't hit the issue unless the
install is cancelled. However, when we start pre-downloading snaps we
would start hitting this in every delayed auto-refresh (that's how I noticed it).

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>