store: decide cache hit/miss at cache level
Instead of exposing an error from the cache interface with which the
store code then decided if the cache was hit or not, make that decision
at the cache level. Changes the cache interface accordingly so the error
is not exposed and the result of a Get is just true/false.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>