multiple: Add missing updating of snippets when changing the homedirs.
Move the functionality of /var/lib/snapd/system-params to its own package with unit tests, to get a more structured approach to the file.

Let SetupSnapConfineSnippets itself discover the homedirs instead of passing it as a parameter, this is also more in line with how it already functions with existing features.