i/b/fwupd: allow writing in EFI's fwupd directory
fwupd is going to use its own EFI directory to be able to handle
multiple distros from the snap. For that it will need to copy Ubuntu's
signed shim. And to not pollute the distro's directory, it will use
its own. So we should just give it permission to put any file there.