gadget: include OnDiskStructure in LaidOutStructure (#12515)
Include OnDiskStructure in LaidOutStructure so we really have all the
needed information to create partitions and content in
LaidOutStructure. Also, remove the now unneeded
onDiskAndLaidoutStructure.

* gadget: change FindDeviceForStructure input to a VolumeStructure

as we did not need the full LaidOutStructure. This also simplifies the
tests.

* gadget/install: use OnDiskStructure in ensureNodesExist

As we did not need the full laid out structure.

* gadget: include OnDiskStructure in LaidOutStructure

Include OnDiskStructure in LaidOutStructure so we really have all the
needed information to create partitions and content in
LaidOutStructure. Also, remove the now unneeded
onDiskAndLaidoutStructure.

* gadget: remove StartOffset from LaidOutStructure

It is not needed as we have the same field in OnDiskStructure, which
is contained in LaidOutStructure.

* gadget: replace access to LaidOutStructure.VolumeStructure by methods

where possible.

* gadget: set encryption info in gadget.LaidoutVolume

Pass around encryption information so we can set the encryption info
in LaidOutStructure nearer to where it is created, which simplifies
the code and fixes cases not considered before.

* tests/muinstaller: use temporarily commit from branch for the build

To be removed as soon as https://github.com/snapcore/snapd/pull/12515
is merged.