o/snapstate: keep monitoring across restarts
Persist some monitoring-related information so that, once snapd starts
up, new monitoring goroutines can be set up if any existed before a
restart. This logic ensures that the pre-download/continue auto-refresh
UX flow survives restarts. This is important because if snapd is part
of the auto-refresh, it will trigger a restart once it refreshes which
would then wipe the monitoring state.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>