tests: get go-1.18 from the xenial PPA
Debian 10 (buster) does no longer accept backports so we cannot
update golang there. In order to ensure snapd works on buster
we need to build it using a PPA with golang 1.18. This should
be okay as debian will re-exec into the new snapd on real
Debian 10 installs.

However eventually we will need to drop support for Debian 10
because of this limiation but for now it is okay to keep it a
bit longer until Debian 12 is really out.