o/snapstate: fix test race condition
go test -race flagged a race condition between two tests' TearDown()
functions and the notification code that runs in a goroutine. This can
only happen in tests where we mock the root dir, so it's not a problem
in production code. This commit simply mocks the notification function
so the race condition isn't triggered.

Signed-off-by: Miguel Pires <miguel.pires@canonical.com>