go.mod: move to go 1.18 here too (#12719)
* packaging: update for go-1.18

* spread: remove opensuse-15.3-64, EOL was 2022-12-31

* tests: xenial needs the ppa:snappy-dev/image PPA to build now

* tests: on buster (Debian 10) only golang 1.19 is in -backports so use that

And only in buster-backports-sloppy right now

* tests: get go-1.18 from the xenial PPA

Debian 10 (buster) does no longer accept backports so we cannot
update golang there. In order to ensure snapd works on buster
we need to build it using a PPA with golang 1.18. This should
be okay as debian will re-exec into the new snapd on real
Debian 10 installs.

However eventually we will need to drop support for Debian 10
because of this limiation but for now it is okay to keep it a
bit longer until Debian 12 is really out.

* go: move go.mod to 1.18 and run `go mod {vendor,tidy}`

* workflow: get vendored dependencies in macos test

* workflow: get deps before mkversion.sh runs in "Build C" step

* vendor: drop including vendor/modules.txt in git again

* tests: get deps before installing the fakestore code (needed on debian-sid)

* Revert "tests: get deps before installing the fakestore code (needed on debian-sid)"

This reverts commit 8f215ee890de4aa401558b19847f07d358993358.

* tests: build test helpers before removing the vendoring

* tests: fix debian-sid build

* Revert "tests: build test helpers before removing the vendoring"

This reverts commit e7d8ba53a887955f573e1ddcac7e7e6beaea6585.