prompting/storage: fixed typo in json name string
Currently there are three different names related to each of the maps.

- `labelDB.Allow`: `json:"allow"` (no corresponding `extras` name)
- `labelDB.AllowWithDir`: `json:"allow-with-dir"`, `extras: "allow-directory"`
- `labelDB.AllowWithSubdirs`: `json:"allow-with-subdir"`, `extras: "allow-subdirectories"`

This is becoming rather untenable.  For example, lack of `s` at end of
`"allow-with-subdir"` is inconsistent with the other two, hence this
bug.

I think it is important to unify the json and extras names, and then
adjust the `labelDB` field names to match.

Note: any changes to json names will *not* be backwards compatible with
existing `/var/lib/snapd/prompt.json` file.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>