prompting/storage: improved handling of paths
Firstly, this change causes "allow-subdirectories" to have the same
effect as "allow-directory" with respect to storing the parent of the
path instead of the path itself.  This changed the behavior of some of
the existing tests, which have been adjusted to explicitly treat paths
which should be directories as directories (if they were on a real
filesystem) by appending a trailing `/` to them.

In general, it is now possible to force any path to be treated by
`Set()` as a directory by appending a trailing `/` to it.

Secondly, this change adds cleaning of paths using `filepath.Clean()`
and calling `filepath.Dir(path)` if the path has a trailing `/` (in
addition to the existing checks based on the state of `extras`).  This
intends to guarantee that if a path should match, it will, and that no
paths in the database should ever have a trailing `/`.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>