many: add pointer to volume in gadget.VolumeStructure.
With this pointer we can access to "partial" options from the Volume,
with information that we need to check if the structure needs a
filesystem or if it has a partially defined size.

Additionally, rename willHaveFilesystem() to HasFilesystem() and use
it everywhere, as in the end they mean the same: the structure
definition in the gadget expects a filesystem to be there.