prompting: modified `notifier.Request` and `storage.StoredDecision`
Added `Snap` and `App` information to requests so that prompt clients
do not have to parse apparmor labels, and because that is used in the
`StoredDecision` struct.  Also added `ResourceType` to the request so
that this information can be displayed in the prompt to the user.  The
`newRequest()` function has been modified to parse the label from
apparmor into `Snap` and `App`, and to assign a `ResourceType` of
`"file"` if `msg.Class == apparmor.MediationClassFile` (which is the
only handled case in `newRequest()` at the moment).

Also, a `newStoredDecision()` function has been added which constructs a
new `StoredDecision` struct based on the information from a given
`notifier.Request` and other information which is produced early on
during the `Set()` function.  The purpose of `newStoredDecision()` is to
be a helper function for `Set()` which create decisions with unique IDs
and sets up most of the boilerplate for the decisions.  The
`Permissions` field is left as an empty list to which `Set()` will
append each permission which is actually added to the DB, since any
which are implied by previous decisions are not added.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>