prompting/storage: return original state of modified decisions
Previously, only the IDs of modified and deleted decisions were returned
by functions which caused those modifications (`Set()` being the primary
exported one).  These modifications and deletions can occur as the
result of pruning when a new decision overrides a previous.  However,
since stored decisions are modified immediately during pruning when a
corresponding leaf is removed from the allow tree, and the `ById` map is
immediately updated to reflect that change, before the end of the
overarching function call which results in the changes, the original
states of the modified decisions have been list, and any deleted
decisions have been removed from the storage entirely.

This commit changes that.  Now, whenever a stored decision is modified
for the first time as a result of pruning, its original state is deep
copied and stored so that it can be later returned to the caller.

As this is more complicated than simply returning the IDs of changed
decisions, some refactoring was done in an attempt to reduce duplicate
code and increase modularity.  Some helper functions were also exported
for testing.

Still TODO: now that the original state of changes decisions are stored,
use that to roll back changes to other decisions if/when an error
occurs.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>