o/ifacestate/apparmorprompting: issue notices for requests and rules
When a new request or rule is created, modified, deleted, replied to,
etc., issue a notice with that that request/rules's ID. Rather than pass
a pointer to the current state into the `promptrequests` and
`accessrules` subsystems, however, the `apparmorprompting` manager
creates functions which ensure that the respective subsystems create the
correct notice types and do not otherwise have access to state.

Currently, the notices which are created have empty `AddNoticeOptions`,
so do not include any data about the state of the relevant request or
rule. It is the responsibility of the client to query the prompting API
endpoints to get the current state of requests/rules. That API is the
single source of truth.

However, since request/rule deletion (because a request is answered or
dropped, because a rule expires or is deleted, etc.) cannot be un-done,
it may be worth including a `"deleted"` field in the notice data if it
is known that a request or rule was deleted. That way the client would
know that the request/rule was deleted without having to query the
prompting API. The one exception where a notice could contain the
`"deleted"` field without the rule actually being deleted is if
`RefreshTreeEnforceConsistency` causes a rule with the same ID to be
re-added due to some inconsistency with the rules or because the rules
were modified without saving and then re-loaded from the disk. These
scenarios should not happen in practice.

Additionally, some small improvements were made to `accessrules`.
Namely, rule expiration checking was moved to its own method on the
`AccessRule` type, and expired rules are now pruned by the `Rules*`
methods as well, rather than being left in the DB until a path is
matched against them. Additionally, those `Rules*` functions now share a
common helper function, since the helper function now handles pruning as
well, and pass a filter function into the helper function.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>