s/apparmor/notify/listener: add interface list to listener request
For now, this is simply a placeholder. TODO: set the interface correctly
using the tag in the message received from the kernel.

It is useful for the other prompting components to know which interfaces
were associated with the rule which triggered a particular request, since
the permissions, path handling, or other properties vary based on the
interface.

Multiple rules from different interfaces may have matched the same
request. In this case, the kernel sends back all the interfaces, and the
listener parses these into a list and includes it in the listener
request. If there is only one interface in the kernel message, the
listener includes a list containing that one interface.

We want to settle on a single interface for each prompting request and
rule. However, it is not the responsibility of the listener to decide
which interface has priority, nor to be aware of interfaces in general;
the listener should simply parse the tag in the kernel message, and pass
the interfaces on to the other prompting components to decide which has
priority.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>