o/i/apparmorprompting: send back errors when prompting not enabled
Whether or not prompting is enabled, API requests to
`/v2/interfaces/prompting/{requests,rules}` make calls to the prompting
subsystem, since the daemon should not know about prompting in
particular or care if it is enabled. However, these calls attempted to
dereference the requests and rules instances, which are nil if prompting
is not enabled. This causes a panic.

If AppArmor Prompting is enabled, the promptrequests and accessrules
instances should be non-nil. Thus, by checking if prompting is enabled
and returning an error if not, these nil-pointer dereferences can be
avoided.

If prompting is enabled and the promptrequests or accessrules instances
are nonetheless nil, there is a real bug which should be addressed, so a
panic is desirable in this case for the stack trace.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>