many: renamed prompting components according to spec changes
Renamed `AccessRule` to `Rule`, and `PromptRequest` to `Prompt`.

This has the convenient side effect of disambiguating listener requests
from prompts (formerly "prompt requests"). Both rules and prompts are
under the "interfaces/requests" API path, but the "requests" prefix can
be omitted aside of the API, notice types, and rule filename.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>