o/i/a/requestprompts: modify outstanding prompt permissions
When a new rule is created, outstanding prompts are checked against it
to see if any are satisfied by it. Previously, if the prompt wasn't
fully satisfied, it was left unmodified, even if there were rules which
covered some (but not all) of its permissions. This is confusing.

Now, if any permissions are satisfied by a new allow rule, those
permissions are removed from the outstanding prompt, causing a new
notice to be issued. If any permissions are matched by a new deny rule,
then that outstanding prompt is immediately satisfied with a deny reply.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>