o/i/a/common: adjust precedence and allow greater wildcard flexibility
Precedence has been adjusted so that *where* has priority over *what*.
That is, base path length takes priority over the specificity of file
extensions or other suffixes. If two patterns have the same base path,
but one follows with `/*` while the another follows with `/**`, the
former has precedence, but if a base+`/*` rule has a shorter base path
than a base+`/**` rule, the latter has precedence.

For example, the pattern `/foo/bar/**` has precedence over /foo/bar*`,
since the former requires that the path ends in `/` (TODO: this isn't
actually the case yet, since path matching hasn't yet been updated to be
more strict now that paths and path patterns may end in `/` and groups
are allowed in path patterns). Both of these patterns match `/foo/bar/`
(and also, currently, incorrectly, `/foo/bar` -- this will be fixed).

Wildcards are no longer restricted to being at the end of the path
pattern, grouped together and followed by a file extension or nothing at
all. Now, characters and/or directories can appear arbitrarily around
the single `*`, and after the `/**/`. However, there must still be at
most one `/**/` and one `*` in the pattern, with the `*` occurring after
the `/**/` if it exists.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>