o/i/a/common: allow arbitrary wildcards in path patterns
There are no longer restrictions about where wildcards or doublestars
may occur in path patterns. Prior modifications to precedence
computation which placed priority on *where* instead of *what* mean that
precedence between any two patterns (without groups or classes) which
match the same path can be decided.

More broadly, the concept of suffixes is no longer used in precedence
computation (due to the previous change) or checking the validity of
path patterns (new in this commit).

As such, the allowable path patterns glob can be modified to be far more
flexible, with wildcards occurring anywhere and the group beginning
anywhere. Additionally, we now allow "mis-use" of `**`, such as
concatenated to a non-`/` character, which is equivalent to `*`. Since
the group may now begin anywhere, we also allow repeated `//`, which is
equivalent to `/`, since it is difficult to check for these prior to
expansion. In both cases, `ExpandPathPattern()` replaces these repeated
characters with a single `*` or `/`, respectively, so that conflicts
with other equivalent patterns still occur as usual.

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>