o/i/a/common: fixed path pattern precedence
Re-architected the pattern precedence system, increasing performance,
clarity, and correctness. Notably, adjusted the precedence of patterns
with a glob as the next character compared to other patterns with a path
separator as the next character.

The new precedence order is as follows (highest to lowest precedence):

literals:
- /foo/bar/baz
- /foo/bar/
terminated:
- /foo/bar
singlestars:
- /foo/bar/*/baz
- /foo/bar/*/
- /foo/bar/*/*baz
- /foo/bar/*/*
- /foo/bar/*
- /foo/bar/*/**
glob:
- /foo/bar*baz
- /foo/bar*/baz
- /foo/bar*/baz/**
- /foo/bar*/
- /foo/bar*/*baz
- /foo/bar*/*/baz
- /foo/bar*/*/
- /foo/bar*/*
- /foo/bar*
doublestars:
- /foo/bar/**/baz
- /foo/bar/**/*baz/
- /foo/bar/**/*baz
terminal doublestar:
- /foo/bar/**/        # These are tough... usually, /foo/bar/**/ would have precedence over
- /foo/bar/**/*       # precedence over /foo/bar/**/*baz, but in this case,
- /foo/bar/**         # the trailing *baz adds more specificity.
glob with immediate doublestar:
- /foo/bar*/**/baz
- /foo/bar*/**/
- /foo/bar*/**

Signed-off-by: Oliver Calder <oliver.calder@canonical.com>